Add sample code
Major application and database overhaul

This commit represents a significant overhaul of the FriendzyBlog application, focusing on enhancing modularity, infrastructure, logging capabilities, and refining the domain model. Key changes include the integration of the MuonroiBuildingBlock module, restructuring of the API's endpoint architecture, and a strategic shift in project dependencies and configurations. Notably, the removal of the HomeController class and Swagger/OpenAPI setup indicates a simplification of the API layer. The project files have seen substantial updates, including the addition of new project references, package references, and a reorganization of the project structure to accommodate these changes.

Significant modifications were made to the database schema and management, including the removal of initial migrations and the entire database context, signaling a pivot in the database strategy or a preparation for a new schema introduction. The addition of Elasticsearch and Kibana configurations, along with Docker Compose setups for these services, suggests an enhancement in the application's infrastructure for better monitoring and logging capabilities.

Furthermore, the application's bootstrapping process in Program.cs has been overhauled to align with the new modular and scalable architecture. Adjustments to global usings reflect the changes in project dependencies and a move towards a more integrated and efficient development workflow.

Lastly, the introduction of structured data seeding processes for initial languages, roles, and users prepares the application for first-time use and ensures essential data is present, highlighting a thoughtful approach to application setup and security.

Overall, this commit marks a strategic enhancement of the FriendzyBlog application, laying a solid foundation for future development and scalability.